# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "BackEnd API",
        "title": "Backend API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/movies": {
            "get": {
                "description": "Get All Movie",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Movie"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Search Movies",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page Movies",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Limit Movies",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort Movies",
                        "name": "sort",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/controllers.TaskResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "result": {
                                            "$ref": "#/definitions/models.Allmovie"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/movies/addmovie": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Create New Movie",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Movie"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Title",
                        "name": "title",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Image Movies",
                        "name": "image_movie",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Genre",
                        "name": "genre",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Release Date",
                        "name": "release_date",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Duration",
                        "name": "duration",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Director",
                        "name": "director",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Cst Actor",
                        "name": "cast_actor",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Synopsis",
                        "name": "synopsis",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/controllers.TaskResponse2"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "result": {
                                            "$ref": "#/definitions/models.Movie_body"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/movies/{id}": {
            "get": {
                "description": "Get Detail Movie",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Movie"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Detail Movies",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/controllers.TaskResponse2"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "results": {
                                            "$ref": "#/definitions/models.Movie_cinema"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Delete Movie",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Movie"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Id Movies",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/controllers.TaskResponse2"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "result": {
                                            "$ref": "#/definitions/models.Movie"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Update Movie",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Movie"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Id Movies",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Title",
                        "name": "title",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Image Movies",
                        "name": "image_movie",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Genre",
                        "name": "genre",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Release Date",
                        "name": "release_date",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Duration",
                        "name": "duration",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Director",
                        "name": "director",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Cst Actor",
                        "name": "cast_actor",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Synopsis",
                        "name": "synopsis",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/controllers.TaskResponse2"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "result": {
                                            "$ref": "#/definitions/models.Movie_body"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "controllers.PageInfo": {
            "type": "object",
            "properties": {
                "currentPage": {
                    "type": "integer",
                    "example": 1
                },
                "nextPage": {
                    "type": "integer",
                    "example": 102
                },
                "prevPage": {
                    "type": "integer",
                    "example": 0
                },
                "totalData": {
                    "type": "integer",
                    "example": 101
                },
                "totalPage": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "controllers.TaskResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "pageInfo": {
                    "$ref": "#/definitions/controllers.PageInfo"
                },
                "result": {},
                "success": {
                    "type": "boolean"
                }
            }
        },
        "controllers.TaskResponse2": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "result": {},
                "success": {
                    "type": "boolean"
                }
            }
        },
        "models.Allmovie": {
            "type": "object",
            "properties": {
                "genre": {
                    "type": "string",
                    "example": "Action, Advanture"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "image_movie": {
                    "type": "string",
                    "example": "b00db012-1a27-43b3-895a-abd3f540362e.jpg"
                },
                "title": {
                    "type": "string",
                    "example": "Spiderman"
                }
            }
        },
...
        "models.Movie": {
            "type": "object",
            "properties": {
                "cast_actor": {
                    "type": "string",
                    "example": "warsito"
                },
                "director": {
                    "type": "string",
                    "example": "joko anwar"
                },
                "duration": {
                    "type": "string",
                    "example": "22:10:33"
                },
                "genre": {
                    "type": "string",
                    "example": "Action, Advanture"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "image_movie": {
                    "type": "string",
                    "example": "b00db012-1a27-43b3-895a-abd3f540362e.jpg"
                },
                "release_date": {
                    "type": "string",
                    "example": "2006-01-02"
                },
                "synopsis": {
                    "type": "string",
                    "example": "Lorem ipsum dolor sit amet"
                },
                "title": {
                    "type": "string",
                    "example": "Spiderman"
                }
            }
        },
        "models.Movie_body": {
            "type": "object",
            "properties": {
                "cast_actor": {
                    "type": "string",
                    "example": "warsito"
                },
                "director": {
                    "type": "string",
                    "example": "joko anwar"
                },
                "duration": {
                    "type": "string",
                    "example": "22:10:33"
                },
                "genre": {
                    "type": "string",
                    "example": "Action, Advanture"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "image_movie": {
                    "type": "string",
                    "example": "b00db012-1a27-43b3-895a-abd3f540362e.jpg"
                },
                "release_date": {
                    "description": "Id           int    `json:\"id\" `\nTitle        string `json:\"title\" form:\"title\"`\nImage_movie  string `json:\"image_movie\" form:\"image_movie\" `\nGenre        string `json:\"genre\" form:\"genre\" `",
                    "type": "string",
                    "example": "2006-01-02"
                },
                "synopsis": {
                    "type": "string",
                    "example": "Lorem ipsum dolor sit amet"
                },
                "title": {
                    "type": "string",
                    "example": "Spiderman"
                }
            }
        },
        "models.Movie_cinema": {
            "type": "object",
            "properties": {
                "cast_actor": {
                    "type": "string",
                    "example": "Tom Holland, Andrew Garfield, Tobey Maguire, Zendaya"
                },
                "director": {
                    "type": "string",
                    "example": "Steven Spielberg"
                },
                "duration": {
                    "type": "string",
                    "example": "02:09:09"
                },
                "genre": {
                    "type": "string",
                    "example": "Action, Advanture"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "image_movie": {
                    "type": "string",
                    "example": "b00db012-1a27-43b3-895a-abd3f540362e.jpg"
                },
                "release_date": {
                    "type": "string",
                    "example": "2025-10-10"
                },
                "synopsis": {
                    "type": "string",
                    "example": "Lorem ipsum dolor sit amet"
                },
                "title": {
                    "type": "string",
                    "example": "Spiderman"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}